requirements.txt
aio-pika==9.5.4
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiormq==6.8.1
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.2.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
openai==1.57.2
packaging==24.2
pamqp==3.3.0
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
python-telegram-bot==21.9
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
